Display project users' common name rather than uid
We are used to looking at common names (wikitech user names) and not
uids (shell names) in lists of project members. Do this mapping by
directly querying LDAP instead of asking keystone for the details of the
user account. Keystone can only fetch a single user at a time (RESTful
APIs are not always awesome) and that takes more than a minute for
a project like tools that has a very large number of users.