Unwrap LDAP attributes
The LDAP result API returns all attributes as lists. This is nice for
consistency when fetching things that may or may not have multiple
values, but its annoying to deal with in the template layer for things
like uid and cn that we know will always have a single value.