Add server detail page
Mimic the server detail page on wikitech using the data available via
the nova API. Not all data posted to wikitech is available via the
anonymous API. Notably missing is the OpenStack host that is actually
hosting the image. This is only available when using an admin login to
the API server. We also do not get the security groups, region name, and
filled quota.