Add global usage stats
Compute usage stats on:
* number of projects
* number of instances
* total ram in use by instances
* total storage quota for instances
* total cpus in use by instances
* number of labs use accounts

These stats are fairly heavy to compute (~1 minute best case), so they
will be cached in redis for 25 hours. A cron job will be created in the
tool's account to update the statistics before they expire. Updating
once per day is probably enough for now. The numbers are not expected to
change radically from day to day.