Use paged LDAP queries for everything
Our OpenLDAP servers are setup with limits on the number of results that
an unprivledged user can get for any given query. The limit at the time
I'm writing this patch is 2048 records. WE dug around and found that it
is possible to set different limits for LDAP queries which use the
'pagedResultsControl' LDAP extension.

* https://ldap3.readthedocs.io/searches.html#simple-paged-search
* http://www.openldap.org/doc/admin24/limits.html#Size%20limits%20and%20Paged%20Results