Add some hiera key browsing
This adds /hierakey/<key> urls to show all users of a given
hiera key.  It also displays the hiera config on each
server page.

I didn't add a top-level 'every hiera key' page -- I'm
not sure if that's useful or not.  Missing it
does make the breadcrumbs a bit weird...