nova: add a region lookup to everything
If this is anything like nfs-exportd, then the addition of a new region
requires every query to be scoped by region when a client is opened.

In nfs-exportd, there was a funny little dance required to get the regions,
auth-wise, but we could always see how this acts.  Right now, it's simply
broken.