Procfile: increase gunicorn timeout to 5 minutes
Pulling all of the information from the OpenStack API that is needed to
populate the statistics table on the app's landing page take quite
a while. Sometimes 60 seconds is long enough, but seemingly only if
there have been enough recent attempts that have warmed up OpenStack's
answers. This slowness is why the app tries to cache the stats for 25
hours and also has a cron job to refresh that cache once per day.

When we were using uwsgi to run the app folks seemed to notice the
landing page 502's a lot less. We did not have any explicit timeouts in
the uwsgi configuration. Let's try a 5 minute (300 second) with gunicorn
to see if that helps.

Bug: T337265